# public/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "IHttpResponse": {
        "properties": {
          "data": {},
          "message": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["data", "message", "status"],
        "type": "object"
      },
      "IUpdateUser": {
        "properties": {
          "photo": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["photo", "name"],
        "type": "object"
      },
      "ICreateUser": {
        "properties": {
          "role": {
            "type": "string",
            "enum": ["admin", "user"]
          },
          "photo": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["photo", "password", "email", "name"],
        "type": "object"
      },
      "ILoginUser": {
        "properties": {
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": ["password", "email"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      },
      "tsoa_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://swagger.io/api/oauth/dialog",
            "scopes": {
              "write:pets": "modify things",
              "read:pets": "read things"
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "tdd-node",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "get": {
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [],
        "parameters": []
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "GetOneUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "operationId": "UpdateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IUpdateUser"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [
          {
            "bearer": ["admin"]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/users/register": {
      "post": {
        "operationId": "RegisterUser",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "operationId": "SignIn",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ILoginUser"
              }
            }
          }
        }
      }
    },
    "/api/v1/users/send-email": {
      "get": {
        "operationId": "SendEmail",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpResponse"
                }
              }
            }
          }
        },
        "tags": ["Users"],
        "security": [],
        "parameters": []
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}